requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
Pillow==10.4.0
python-decouple==3.8
coreapi==2.3.3
pyyaml==6.0.2 